@@ -2,4 +2,4 @@ from .cfg.config import set_coordinate_branch
 from .io.geo_handler import load_geometry
 from .vis.geo import plot_geometry
 
-__all__ = ["pygeosimplify", "load_geometry", "plot_geometry", "set_coordinate_branch"]
+__all__ = ["load_geometry", "plot_geometry", "pygeosimplify", "set_coordinate_branch"]
